remove labse, replace with sentencetransformer; update example
