+++ bertalign-rs/src/embed/utils.rs
@@ -1,10 +1,10 @@
-use candle_core::{DType, Device};
+use candle_core::{DType, Device, Tensor};
 use candle_nn::var_builder::{SimpleBackend, VarBuilderArgs};
 use hf_hub::api::sync::Api;
 
 use std::path::PathBuf;
 
-use crate::error::{DownloadHFModelError, LoadSafeTensorError};
+use crate::error::{DownloadHFModelError, FastTokenBatchError, LoadSafeTensorError};
 
 /// memmap is unsafe because the file can be modified during or after
 /// the file is read.
@@ -22,4 +22,127 @@ pub fn download_hf_model(model_id: &str, filename: &str) -> Result<PathBuf, Down
     let model_repo = api.model(model_id.to_string());
     let filepath = model_repo.get(filename)?;
     Ok(filepath)
+}
+
+/// Below is the code for doing token-based batching
+/// Given input_ids, and some `batch_size`, we build batches such that
+/// the max number of tokens doesn't exceed `batch_size`. We also make
+/// sure that batches are in multiples of 8 for better GPU utilization.
+///
+/// The TokenBatchOutput has the original_ids field, which contains the
+/// original location of the sentences before batching as well as the
+/// sentence length which was used for sorted and batching.
+///
+/// Will refactor later.
+pub struct Batch {
+    pub input_ids: Tensor,
+    pub attention_mask: Tensor,
+    pub token_type_ids: Tensor,
+}
+
+pub struct TokenBatchOutput {
+    pub batches: Vec<Batch>,
+    pub original_ids: Vec<(usize, usize)>, // this is currently the idx and the sentence len / token cnt
+}
+
+pub fn fast_token_based_batching(
+    token_ids: &mut Vec<Vec<u32>>,
+    batch_size: usize,
+    device: &Device,
+) -> Result<TokenBatchOutput, FastTokenBatchError> {
+    let mut finalized_batches = vec![];
+
+    // sort the sentences based on length as if it was already paded to the first multiple-of-8th token
+    let mut sentence_lens = token_ids
+        .iter()
+        .enumerate()
+        .map(|(i, t)| {
+            let tok_len = 8 * ((t.len() + 7) / 8);
+            (i, tok_len)
+        })
+        .collect::<Vec<(usize, usize)>>();
+    sentence_lens.sort_by_key(|(_, len)| *len);
+
+    // collate the data into batches that fill the batch_size
+    let mut batches: Vec<Vec<usize>> = vec![];
+    let mut current_batch: Vec<usize> = vec![];
+    let mut current_token_count = 0;
+
+    for (idx, tok_len) in sentence_lens.iter() {
+        let idx = *idx;
+        let tok_len = *tok_len;
+
+        if current_token_count + tok_len > batch_size {
+            batches.push(std::mem::take(&mut current_batch));
+            current_batch.push(idx);
+            current_token_count = tok_len;
+        } else {
+            current_batch.push(idx);
+            current_token_count += tok_len;
+        }
+    }
+    if current_batch.len() > 0 {
+        batches.push(std::mem::take(&mut current_batch));
+    }
+
+    // for each batch run embeddings
+    for b in batches.iter() {
+        let mut current_token_ids = b
+            .iter()
+            .map(|idx| std::mem::take(&mut token_ids[*idx]))
+            .collect::<Vec<Vec<u32>>>();
+
+        let mut max_seq_length = current_token_ids.iter().map(|tok| tok.len()).max().unwrap();
+        max_seq_length = 8 * ((max_seq_length + 7) / 8);
+
+        let mut current_attn_mask = vec![];
+
+        // pad tokens
+        for i in 0..current_token_ids.len() {
+            // attention mask first, since it depends on len of input_ids before padding
+            let mut this_attn_mask = vec![1u32; current_token_ids[i].len()];
+            this_attn_mask.resize(max_seq_length, 0u32);
+            current_attn_mask.push(std::mem::take(&mut this_attn_mask));
+
+            // pad token_ids
+            current_token_ids[i].resize(max_seq_length, 0u32);
+        }
+
+        // convert input_ids to tensor
+        let (a, b) = (current_token_ids.len(), max_seq_length);
+        let current_token_ids = Tensor::from_vec(
+            current_token_ids
+                .into_iter()
+                .flatten()
+                .collect::<Vec<u32>>(),
+            a * b,
+            device,
+        )?
+        .reshape((a, b))?;
+
+        // convert attn_mask to tensor
+        let current_attn_mask = Tensor::from_vec(
+            current_attn_mask
+                .into_iter()
+                .flatten()
+                .collect::<Vec<u32>>(),
+            a * b,
+            device,
+        )?
+        .reshape((a, b))?;
+
+        // create token_type_ids
+        let current_token_type_ids = Tensor::zeros_like(&current_token_ids)?;
+
+        finalized_batches.push(Batch {
+            input_ids: current_token_ids,
+            attention_mask: current_attn_mask,
+            token_type_ids: current_token_type_ids,
+        });
+    }
+
+    Ok(TokenBatchOutput {
+        batches: finalized_batches,
+        original_ids: sentence_lens,
+    })
 }
